Share EFCore tests code
