+++ net/DevExtreme.AspNet.Data.Tests.EFCore/PaginateViaPrimaryKey.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace DevExtreme.AspNet.Data.Tests.EFCore2 {
+namespace DevExtreme.AspNet.Data.Tests.EFCore {
 
     public class PaginateViaPrimaryKey {
