+++ net/DevExtreme.AspNet.Data.Tests.EFCore/RemoteGroupingStress.cs
@@ -1,9 +1,10 @@
-﻿using System;
+﻿#if !EFCORE1
+using System;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace DevExtreme.AspNet.Data.Tests.EFCore2 {
+namespace DevExtreme.AspNet.Data.Tests.EFCore {
 
     public class RemoteGroupingStress {
 
@@ -31,3 +32,4 @@ namespace DevExtreme.AspNet.Data.Tests.EFCore2 {
     }
 
 }
+#endif
