+++ net/DevExtreme.AspNet.Data.Tests.EFCore1/DevExtreme.AspNet.Data.Tests.EFCore1.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>    
     <TargetFramework>netcoreapp2.0</TargetFramework>
+    <DefineConstants>EFCORE1</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
@@ -16,6 +17,10 @@
   <ItemGroup>
     <ProjectReference Include="..\DevExtreme.AspNet.Data\DevExtreme.AspNet.Data.csproj" />
     <ProjectReference Include="..\DevExtreme.AspNet.Data.Tests.Common\DevExtreme.AspNet.Data.Tests.Common.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\DevExtreme.AspNet.Data.Tests.EFCore\**\*.cs" />
   </ItemGroup>
 
 </Project>
