--- net/DevExtreme.AspNet.Data.Tests.EFCore1/TestDbContext.cs
@@ -1,38 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-
-namespace DevExtreme.AspNet.Data.Tests.EFCore1 {
-
-    class TestDbContext : DbContext {
-        static TestDbContext INSTANCE;
-
-        private TestDbContext(DbContextOptions options)
-            : base(options) {
-        }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder) {
-            modelBuilder.Entity<Bug120.DataItem>();
-            modelBuilder.Entity<RemoteGrouping.DataItem>();
-            modelBuilder.Entity<Summary.DataItem>();
-        }
-
-        public static void Exec(Action<TestDbContext> action) {
-            if(INSTANCE == null) {
-                var helper = new SqlServerTestDbHelper("EFCore1");
-                helper.ResetDatabase();
-
-                var options = new DbContextOptionsBuilder()
-                    .UseSqlServer(helper.ConnectionString)
-                    .Options;
-
-                INSTANCE = new TestDbContext(options);
-                INSTANCE.Database.EnsureCreated();
-            }
-
-            action(INSTANCE);
-        }
-
-    }
-
-}
