--- net/DevExtreme.AspNet.Data.Tests.EFCore2/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿using Xunit;
-
-[assembly: CollectionBehavior(DisableTestParallelization = true)]
