+++ net/DevExtreme.AspNet.Data.Tests.EFCore2/DevExtreme.AspNet.Data.Tests.EFCore2.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.0</TargetFramework>
+    <DefineConstants>EFCORE2</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
@@ -14,6 +15,10 @@
   <ItemGroup>
     <ProjectReference Include="..\DevExtreme.AspNet.Data\DevExtreme.AspNet.Data.csproj" />
     <ProjectReference Include="..\DevExtreme.AspNet.Data.Tests.Common\DevExtreme.AspNet.Data.Tests.Common.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\DevExtreme.AspNet.Data.Tests.EFCore\**\*.cs" />
   </ItemGroup>
 
 </Project>
